+++ readme.php
@@ -22,6 +22,9 @@ class User
     /** @var string */
     public $name;
 
+    /** @var int */
+    public $alternativeId;
+
     public function greet()
     {
         echo 'Hello, ' . $this->name . '!';
@@ -57,4 +60,51 @@ class Role
 
     /** @var bool */
     public $invisible;
+}
+
+class AgendaLine
+{
+    use Neat\Object\Storage;
+
+    /** @var int */
+    public $id;
+
+    /** @var int */
+    public $appointmentId;
+
+    /** @var string */
+    public $description;
+}
+
+class Appointment
+{
+    use Neat\Object\Storage;
+    use Neat\Object\Relations;
+
+    /** @var int */
+    public $id;
+
+    /** @var int */
+    public $createdBy;
+
+    public function creator(): Neat\Object\Relations\One
+    {
+        return $this->belongsToOne(User::class, 'creator', function (Neat\Object\Relations\Reference\LocalKeyBuilder $builder) {
+            // ...
+        });
+    }
+
+    public function agendaLines(): Neat\Object\Relations\Many
+    {
+        return $this->hasMany(AgendaLine::class, 'agenda', function (Neat\Object\Relations\Reference\RemoteKeyBuilder $builder) {
+            // ...
+        });
+    }
+
+    public function attendees(): Neat\Object\Relations\Many
+    {
+        return $this->belongsToMany(User::class, 'attendees', function (Neat\Object\Relations\Reference\JunctionTableBuilder $builder) {
+            // ...
+        });
+    }
 }