+++ changelog.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Events->triggerIfExists().
 - EventDispatcher->triggerIfExists().
 - Collection / Collectible documentation.
+- Relation $configure parameter of type callable(ReferenceBuilder).
+- Support for setting local and remote keys using property or column names as string.
+
+### Deprecated
+- ReferenceBuilder setRemoteKeyString() in favour of setRemoteKeyColumn().
+- setJunctionTableRemoteForeignKey() in favour of setJunctionTableRemoteKeyColumn().
+- setJunctionTableLocalForeignKey() in favour of setJunctionTableLocalKeyColumn().
+- build*() methods in the Relations trait and ReferenceFactory trait.
+- RelationBuilder->referenceFactory() in favour of $configuration parameter with Relations trait methods.
+- Passing a Property instance to setLocalKey and setRemoteKey is deprecated.
 
 ### Changed
 - EventDispatcher->trigger() now throws if an event is not defined.